Live reload, browserify and test socket io.
A lot of experimentation with the build system. Tried require js, but
couldn't get it to work, then found that browserify supports the same
module interface as node which will be really convenient for
implementing the common game logic.

Got live reload working using the chrome extension.

Tried out socket io to make sure it works with browserify.